+++ src/database/ReplicaIntf.java
@@ -9,6 +9,6 @@ public interface ReplicaIntf extends Remote{
 
 	public boolean keepTransactionAlive(List<LeaseLock> locks) throws RemoteException;
 	
-	public String RWTcommit(Integer transactionID, List<LeaseLock> heldLocks,
+	public String RWTcommit(Long transactionID, List<LeaseLock> heldLocks,
 			HashMap<Integer, Integer> memaddrToValue) throws RemoteException;
 }
